requirements.txt
aiofiles==23.2.1 
aiogram==3.3.0
aiohttp==3.9.2
aiosignal==1.3.1
aiosmtplib==3.0.1
aiosqlite==0.20.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncjson==0.0.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses==0.6
dnspython==1.15.0
dnspython3==1.15.0
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.6
joblib==1.4.0
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.5
nltk==3.8.1
number-parser==0.3.2
openpyxl==3.1.2
phonenumbers==8.13.40
pydantic==2.5.3
pydantic_core==2.14.6
pyderive3==0.0.6
pydns3==0.0.9
pypool3==0.0.2
pyserve3==0.0.7
pystructs3==0.0.7
python-dotenv==1.0.1
python-stdnum==1.20
regex==2024.4.16
requests==2.31.0
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
validate_email==1.3
yarl==1.9.4